requirements.txt
PyQt6>=6.4.0
qasync>=0.23.0
qtawesome>=1.2.3
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.5
gspread>=5.10.0
oauth2client>=4.1.3
rapidfuzz>=3.0.0